[Feature-#1862][s3] Added support for reading compressed format files , writing multiple objects etc.

+++ chunjun-connectors/chunjun-connector-s3/src/main/java/com/dtstack/chunjun/connector/s3/sink/S3OutputFormat.java
@@ -22,6 +22,7 @@ import com.dtstack.chunjun.config.FieldConfig;
 import com.dtstack.chunjun.connector.s3.config.S3Config;
 import com.dtstack.chunjun.connector.s3.util.S3Util;
 import com.dtstack.chunjun.connector.s3.util.WriterUtil;
+import com.dtstack.chunjun.constants.ConstantValue;
 import com.dtstack.chunjun.restore.FormatState;
 import com.dtstack.chunjun.sink.format.BaseRichOutputFormat;
 import com.dtstack.chunjun.throwable.ChunJunRuntimeException;
@@ -32,7 +33,6 @@ import org.apache.flink.table.data.RowData;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PartETag;
-import com.esotericsoftware.minlog.Log;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 
@@ -54,10 +54,10 @@ public class S3OutputFormat extends BaseRichOutputFormat {
     private S3Config s3Config;
 
     /** Must start at 1 and cannot be greater than 10,000 */
-    private static int currentPartNumber;
+    private int currentPartNumber;
 
-    private static String currentUploadId;
-    private static boolean willClose = false;
+    private String currentUploadId;
+    private boolean willClose = false;
     private transient StringWriter sw;
     private transient List<MyPartETag> myPartETags;
 
@@ -67,21 +67,49 @@ public class S3OutputFormat extends BaseRichOutputFormat {
     private static final long MIN_SIZE = 1024 * 1024 * 25L;
 
     @Override
+    public void initializeGlobal(int parallelism) {
+        this.amazonS3 = S3Util.getS3Client(s3Config);
+        if (OVERWRITE_MODE.equalsIgnoreCase(s3Config.getWriteMode())) {
+            checkOutputDir();
+        }
+    }
+
+    @Override
     protected void openInternal(int taskNumber, int numTasks) {
-        openSource();
-        restore();
-        checkOutputDir();
-        createActionFinishedTag();
-        nextBlock();
+        // 写多个对象时
+        if (!s3Config.isWriteSingleObject()) {
+            s3Config.setObject(
+                    s3Config.getObject()
+                            + ConstantValue.SINGLE_SLASH_SYMBOL
+                            + jobId
+                            + "_"
+                            + taskNumber
+                            + getExtension());
+        } else {
+            // 写单个对象时
+            if (OVERWRITE_MODE.equalsIgnoreCase(s3Config.getWriteMode())) {
+                // 当写入模式是overwrite时
+                s3Config.setObject(s3Config.getObject() + getExtension());
+            } else {
+                // 当写入模式是append时
+                s3Config.setObject(
+                        s3Config.getObject() + "_" + jobId + "_" + taskNumber + getExtension());
+            }
+        }
+        log.info("current write object name: {}", s3Config.getObject());
         List<FieldConfig> column = s3Config.getColumn();
         columnNameList = column.stream().map(FieldConfig::getName).collect(Collectors.toList());
         columnTypeList = column.stream().map(FieldConfig::getType).collect(Collectors.toList());
+        openSource();
+        restore();
+        createActionFinishedTag();
+        nextBlock();
     }
 
     private void openSource() {
         this.amazonS3 = S3Util.getS3Client(s3Config);
         this.myPartETags = new ArrayList<>();
-        this.currentPartNumber = taskNumber - numTasks + 1;
+        this.currentPartNumber = 0;
         beforeWriteRecords();
     }
 
@@ -95,17 +123,50 @@ public class S3OutputFormat extends BaseRichOutputFormat {
     }
 
     private void checkOutputDir() {
-        if (S3Util.doesObjectExist(amazonS3, s3Config.getBucket(), s3Config.getObject())) {
-            if (OVERWRITE_MODE.equalsIgnoreCase(s3Config.getWriteMode())) {
-                S3Util.deleteObject(amazonS3, s3Config.getBucket(), s3Config.getObject());
+        // 覆盖写单个对象时
+        if (s3Config.isWriteSingleObject()
+                && S3Util.doesObjectExist(amazonS3, s3Config.getBucket(), s3Config.getObject())) {
+            S3Util.deleteObject(amazonS3, s3Config.getBucket(), s3Config.getObject());
+        }
+        // 覆盖写多个对象时
+        if (!s3Config.isWriteSingleObject()) {
+            List<String> subObjects;
+            if (s3Config.isUseV2()) {
+                subObjects =
+                        S3Util.listObjectsKeyByPrefix(
+                                amazonS3,
+                                s3Config.getBucket(),
+                                s3Config.getObject(),
+                                s3Config.getFetchSize());
+            } else {
+                subObjects =
+                        S3Util.listObjectsByv1(
+                                amazonS3,
+                                s3Config.getBucket(),
+                                s3Config.getObject(),
+                                s3Config.getFetchSize());
             }
+            String[] keys = subObjects.toArray(new String[] {});
+            S3Util.deleteObjects(amazonS3, s3Config.getBucket(), keys);
+            log.info("delete objects num：" + keys.length);
+            log.debug("delete objects list：" + StringUtils.join(keys, ","));
+        }
+    }
+
+    public String getExtension() {
+        if (StringUtils.isNotBlank(s3Config.getSuffix())) {
+            return s3Config.getSuffix();
+        } else {
+            return "";
         }
     }
 
     private void nextBlock() {
-        sw = new StringWriter();
+        if (sw == null) {
+            sw = new StringWriter();
+        }
         this.writerUtil = new WriterUtil(sw, s3Config.getFieldDelimiter());
-        this.currentPartNumber = this.currentPartNumber + numTasks;
+        this.currentPartNumber = this.currentPartNumber + 1;
     }
 
     /** Create file multipart upload ID */
@@ -159,6 +220,7 @@ public class S3OutputFormat extends BaseRichOutputFormat {
                     myPartETags.stream().map(Objects::toString).collect(Collectors.joining(",")));
             writerUtil.close();
             writerUtil = null;
+            sw = null;
         }
     }
 
@@ -192,7 +254,7 @@ public class S3OutputFormat extends BaseRichOutputFormat {
         flushDataInternal();
         completeMultipartUploadFile();
         S3Util.closeS3(amazonS3);
-        Log.info("S3Client close!");
+        log.info("S3Client close!");
     }
 
     @Override
@@ -218,7 +280,7 @@ public class S3OutputFormat extends BaseRichOutputFormat {
             }
             String[] stringRecord = new String[columnNameList.size()];
             // convert row to string
-            rowConverter.toExternal(rowData, stringRecord);
+            stringRecord = (String[]) rowConverter.toExternal(rowData, stringRecord);
             try {
                 for (int i = 0; i < columnNameList.size(); ++i) {
